fix: Set mean adjusted agreement with same variable to NA
#4 Fixed the issue of writing past the end of the vector, however neglected use-cases where `candidates` was not equal to `variables`.
This patch removes setting NA in `mean.index()` entirely, and applies the colname/rowname-wise fix to the created matrix instead.
This fixes the issue where in the README example, the MAA of X7 with X2 would be set to NA, instead of with X7.
This should also keep things stable against shuffled columns, or supplying variables in a different order than candidates, and should not encounter the issue of extra NAs.